Debug option moved to TornadoOptions
