+++ tornado-runtime/src/main/java/uk/ac/manchester/tornado/runtime/common/Tornado.java
@@ -47,7 +47,6 @@ public final class Tornado implements TornadoSettingInterface {
 
     public static final boolean ENABLE_VECTORS = Boolean.parseBoolean(settings.getProperty("tornado.vectors.enable", "True"));
     public static final boolean TORNADO_ENABLE_BIFS = Boolean.parseBoolean(settings.getProperty("tornado.bifs.enable", "False"));
-    public static final boolean DEBUG = Boolean.parseBoolean(settings.getProperty("tornado.debug", "False"));
 
     private static final String TORNADO_SDK_VARIABLE = "TORNADO_SDK";
     public static boolean FORCE_BLOCKING_API_CALLS = false;
